% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setops.R
\name{get_unique_column_set}
\alias{get_unique_column_set}
\title{Get unique sets of items in a target column, grouped by a grouping column.}
\usage{
get_unique_column_set(df, groupcol = "uid", targetcol = "Stripped.Sequence")
}
\description{
Get unique sets of items in a target column, grouped by a grouping column.
}
